requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
huggingface-hub>=0.23.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0